+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 - _Minor change is a one affecting user more or less._
 - Patch change is none of above or fix.
 
-# 18.0.1
+# 18.1.0
 
+- _Do not transfer scroll to window when viewer scroll is not changed._
 - Fix sometimes not appearing F11 failure guide.
 - Navigation 1px precision issue.